+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.types.js
@@ -15,6 +15,7 @@ export type PlainProps = {|
     hideWidgets: HideWidgets,
     teiId: string,
     enrollmentId: string,
+    eventId: string,
     program: Program,
     trackedEntityTypeId: string,
     mode: string,
@@ -30,8 +31,13 @@ export type PlainProps = {|
     onLinkedRecordClick: LinkedRecordClick,
     onEnrollmentError: (message: string) => void,
     onEnrollmentSuccess: () => void,
+    onUpdateEnrollmentStatus: (enrollment: Object) => void,
+    onUpdateEnrollmentStatusSuccess: ({ redirect?: boolean }) => void,
+    onUpdateEnrollmentStatusError: (message: string) => void,
+    onSaveAndCompleteEnrollment: (enrollment: Object) => void,
     onCancelEditEvent: (isScheduled: boolean) => void,
     onHandleScheduleSave: (eventData: Object) => void,
+    onAccessLostFromTransfer?: () => void,
     pageStatus: string,
     eventStatus?: string,
     eventAccess: {|
@@ -42,6 +48,8 @@ export type PlainProps = {|
     assignee: UserFormField | null,
     onSaveAssignee: (newAssignee: UserFormField) => void,
     onSaveAssigneeError: (prevAssignee: UserFormField | null) => void,
+    onDeleteTrackedEntitySuccess: () => void,
+    events: Array<any>,
 |};
 
 export type Props = {|