+++ src/core_modules/capture-core/components/Pages/common/EnrollmentOverviewDomain/index.js
@@ -4,12 +4,16 @@ export {
     enrollmentSiteActionTypes,
     updateEnrollmentDate,
     updateIncidentDate,
-    updateEnrollmentEvents,
+    updateEnrollmentEvent,
     commitEnrollmentEvent,
     rollbackEnrollmentEvent,
-    updateEnrollmentEventsWithoutId,
+    updateEnrollmentEventWithoutId,
     updateEnrollmentAttributeValues,
     showEnrollmentError,
+    updateEnrollmentAndEvents,
+    commitEnrollmentAndEvents,
+    rollbackEnrollmentAndEvents,
+    setExternalEnrollmentStatus,
 } from './enrollment.actions';
 export { useCommonEnrollmentDomainData } from './useCommonEnrollmentDomainData';
 export { useRuleEffects } from './useRuleEffects';