+++ src/core_modules/capture-core/components/WidgetEnrollment/Actions/Transfer/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { Transfer } from './Transfer.component';